+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,0 +1,37 @@
+package roomescape.service;
+
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.domain.ReservationTime;
+import roomescape.controller.dto.ReservationTimeRequest;
+import roomescape.controller.dto.ReservationTimeResponse;
+import roomescape.repository.ReservationDao;
+import roomescape.repository.ReservationTimeDao;
+
+@Service
+public class ReservationTimeService {
+
+    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
+
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
+        this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
+    }
+
+    public List<ReservationTimeResponse> findAll() {
+        List<ReservationTime> reservationTimes = reservationTimeDao.findAll();
+        return reservationTimes.stream()
+                .map(ReservationTimeResponse::from)
+                .toList();
+    }
+
+    public Long create(ReservationTimeRequest reservationTimeRequest) {
+        return reservationTimeDao.insert(reservationTimeRequest.toReservationTime());
+    }
+
+    public void delete(Long id) {
+        reservationDao.deleteByTimeId(id);
+        reservationTimeDao.deleteById(id);
+    }
+}
